refactor!: rename ops::leaf to ops::sum
