+++ hugr-core/src/ops/sum.rs
